Replace students who did not take enough credits at chkpt to null
To prevent data leakage, students that have not reached the 12 credits and not taken the course by the checkpoint term (which this data is limited to at the time of this function), will have the applicable in_12_credits columns set to null. Students that took the course will remain True, regardless of credits achieved.